Add a 'sub_dir' option to deps
This option tells rebar to look for erlang code inside a subdir of the
checked-out repo instead of the repo's root directory.

Example rebar.config usage:
{deps,[{myproject,".*",{git,"git://...",master},[{sub_dir,"erlang"}]}]}

Some code repositories have their erlang code in a subdirectory instead
of the repo's root directory. This is typically the case of projects
that are not erlang-specific, for example because they provide API
implementations in multiple languages. Rebar should adapt to these
projects instead of the other way around.